stylegan-app/src/requirements.txt
absl-py==2.3.1
addict==2.4.0
albucore
albumentations
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
basicsr
beautifulsoup4==4.13.4
build==1.2.2.post1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
easydict==1.13
facexlib
filelock==3.18.0
filterpy
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
gfpgan
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py
humanfriendly==10.0
idna==3.10
imageio
imageio-ffmpeg==0.4.3
insightface
jax
jaxlib
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib
mdurl==0.1.2
mediapipe
ml_dtypes==0.5.1
mpmath==1.3.0
namex
networkx
ninja==1.11.1.4
noise==1.2.2
numba==0.61.2
numpy==2.1.3
onnx==1.18.0
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
prettytable==3.16.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy
sentencepiece
sentry-sdk
simsimd
six
smmap
sounddevice
soupsieve
stringzilla
stylegan2-torch
sympy
tb-nightly
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tifffile
typing-inspection
typing_extensions
urllib3
wandb
wcwidth
Werkzeug
wrapt
yapf
pyspng

